+++ src/bin/jp2/convert.c
@@ -42,15 +42,6 @@
 #include <string.h>
 #include <ctype.h>
 
-#ifdef OPJ_HAVE_LIBTIFF
-#include <tiffio.h>
-#endif /* OPJ_HAVE_LIBTIFF */
-
-#ifdef OPJ_HAVE_LIBPNG
-#include <zlib.h>
-#include <png.h>
-#endif /* OPJ_HAVE_LIBPNG */
-
 #include "openjpeg.h"
 #include "convert.h"
 
@@ -1864,562 +1855,3 @@ int imagetorawl(opj_image_t * image, const char *outfile)
     return imagetoraw_common(image, outfile, OPJ_FALSE);
 }
 
-#ifdef OPJ_HAVE_LIBPNG
-
-#define PNG_MAGIC "\x89PNG\x0d\x0a\x1a\x0a"
-#define MAGIC_SIZE 8
-/* PNG allows bits per sample: 1, 2, 4, 8, 16 */
-
-opj_image_t *pngtoimage(const char *read_idf, opj_cparameters_t * params)
-{
-    png_structp  png;
-    png_infop    info;
-    double gamma, display_exponent;
-    int bit_depth, interlace_type,compression_type, filter_type;
-    int unit;
-    png_uint_32 resx, resy;
-    unsigned int i, j;
-    png_uint_32  width, height;
-    int color_type, has_alpha, is16;
-    unsigned char *s;
-    FILE *reader;
-    unsigned char **rows;
-    /* j2k: */
-    opj_image_t *image;
-    opj_image_cmptparm_t cmptparm[4];
-    int sub_dx, sub_dy;
-    unsigned int nr_comp;
-    int *r, *g, *b, *a = NULL;
-    unsigned char sigbuf[8];
-
-    if((reader = fopen(read_idf, "rb")) == NULL)
-    {
-        fprintf(stderr,"pngtoimage: can not open %s\n",read_idf);
-        return NULL;
-    }
-    image = NULL; png = NULL; rows = NULL;
-
-    if(fread(sigbuf, 1, MAGIC_SIZE, reader) != MAGIC_SIZE
-            || memcmp(sigbuf, PNG_MAGIC, MAGIC_SIZE) != 0)
-    {
-        fprintf(stderr,"pngtoimage: %s is no valid PNG file\n",read_idf);
-        goto fin;
-    }
-    /* libpng-VERSION/example.c:
- * PC : screen_gamma = 2.2;
- * Mac: screen_gamma = 1.7 or 1.0;
-*/
-    display_exponent = 2.2;
-
-    if((png = png_create_read_struct(PNG_LIBPNG_VER_STRING,
-                                     NULL, NULL, NULL)) == NULL)
-        goto fin;
-    if((info = png_create_info_struct(png)) == NULL)
-        goto fin;
-
-    if(setjmp(png_jmpbuf(png)))
-        goto fin;
-
-    png_init_io(png, reader);
-    png_set_sig_bytes(png, MAGIC_SIZE);
-
-    png_read_info(png, info);
-
-    if(png_get_IHDR(png, info, &width, &height,
-                    &bit_depth, &color_type, &interlace_type,
-                    &compression_type, &filter_type) == 0)
-        goto fin;
-
-    /* png_set_expand():
- * expand paletted images to RGB, expand grayscale images of
- * less than 8-bit depth to 8-bit depth, and expand tRNS chunks
- * to alpha channels.
-*/
-    if(color_type == PNG_COLOR_TYPE_PALETTE)
-        png_set_expand(png);
-    else
-        if(color_type == PNG_COLOR_TYPE_GRAY && bit_depth < 8)
-            png_set_expand(png);
-
-    if(png_get_valid(png, info, PNG_INFO_tRNS))
-        png_set_expand(png);
-
-    is16 = (bit_depth == 16);
-
-    /* GRAY => RGB; GRAY_ALPHA => RGBA
-*/
-    if(color_type == PNG_COLOR_TYPE_GRAY
-            || color_type == PNG_COLOR_TYPE_GRAY_ALPHA)
-    {
-        png_set_gray_to_rgb(png);
-        color_type =
-                (color_type == PNG_COLOR_TYPE_GRAY? PNG_COLOR_TYPE_RGB:
-                                                    PNG_COLOR_TYPE_RGB_ALPHA);
-    }
-    if( !png_get_gAMA(png, info, &gamma))
-        gamma = 0.45455;
-
-    png_set_gamma(png, display_exponent, gamma);
-
-    png_read_update_info(png, info);
-
-    png_get_pHYs(png, info, &resx, &resy, &unit);
-
-    color_type = png_get_color_type(png, info);
-
-    has_alpha = (color_type == PNG_COLOR_TYPE_RGB_ALPHA);
-
-    nr_comp = 3 + (unsigned int)has_alpha;
-
-    bit_depth = png_get_bit_depth(png, info);
-
-    rows = (unsigned char**)calloc(height+1, sizeof(unsigned char*));
-    for(i = 0; i < height; ++i)
-        rows[i] = (unsigned char*)malloc(png_get_rowbytes(png,info));
-
-    png_read_image(png, rows);
-
-    memset(cmptparm, 0, sizeof(cmptparm));
-
-    sub_dx = params->subsampling_dx; sub_dy = params->subsampling_dy;
-
-    for(i = 0; i < nr_comp; ++i)
-    {
-        cmptparm[i].prec = (OPJ_UINT32)bit_depth;
-        /* bits_per_pixel: 8 or 16 */
-        cmptparm[i].bpp = (OPJ_UINT32)bit_depth;
-        cmptparm[i].sgnd = 0;
-        cmptparm[i].dx = (OPJ_UINT32)sub_dx;
-        cmptparm[i].dy = (OPJ_UINT32)sub_dy;
-        cmptparm[i].w = (OPJ_UINT32)width;
-        cmptparm[i].h = (OPJ_UINT32)height;
-    }
-
-    image = opj_image_create(nr_comp, &cmptparm[0], OPJ_CLRSPC_SRGB);
-
-    if(image == NULL) goto fin;
-
-    image->x0 = (OPJ_UINT32)params->image_offset_x0;
-    image->y0 = (OPJ_UINT32)params->image_offset_y0;
-    image->x1 = (OPJ_UINT32)(image->x0 + (width  - 1) * (OPJ_UINT32)sub_dx + 1 + image->x0);
-    image->y1 = (OPJ_UINT32)(image->y0 + (height - 1) * (OPJ_UINT32)sub_dy + 1 + image->y0);
-
-    r = image->comps[0].data;
-    g = image->comps[1].data;
-    b = image->comps[2].data;
-    if(has_alpha) {
-        a = image->comps[3].data;
-        image->comps[3].alpha = 1;
-    }
-
-    for(i = 0; i < height; ++i)
-    {
-        s = rows[i];
-
-        for(j = 0; j < width; ++j)
-        {
-            if(is16)
-            {
-                *r++ = s[0]<<8|s[1]; s += 2;
-
-                *g++ = s[0]<<8|s[1]; s += 2;
-
-                *b++ = s[0]<<8|s[1]; s += 2;
-
-                if(has_alpha) { *a++ = s[0]<<8|s[1]; s += 2; }
-
-                continue;
-            }
-            *r++ = *s++; *g++ = *s++; *b++ = *s++;
-
-            if(has_alpha) *a++ = *s++;
-        }
-    }
-fin:
-    if(rows)
-    {
-        for(i = 0; i < height; ++i)
-            free(rows[i]);
-        free(rows);
-    }
-    if(png)
-        png_destroy_read_struct(&png, &info, NULL);
-
-    fclose(reader);
-
-    return image;
-
-}/* pngtoimage() */
-
-int imagetopng(opj_image_t * image, const char *write_idf)
-{
-    FILE *writer;
-    png_structp png;
-    png_infop info;
-    int *red, *green, *blue, *alpha;
-    unsigned char *row_buf, *d;
-    int has_alpha, width, height, nr_comp, color_type;
-    int adjustR, adjustG, adjustB, adjustA, x, y, fails;
-    int prec, ushift, dshift, is16, force16, force8;
-    unsigned short mask = 0xffff;
-    png_color_8 sig_bit;
-
-    is16 = force16 = force8 = ushift = dshift = 0; fails = 1;
-    prec = (int)image->comps[0].prec;
-    nr_comp = (int)image->numcomps;
-
-    if(prec > 8 && prec < 16)
-    {
-        ushift = 16 - prec; dshift = prec - ushift;
-        prec = 16; force16 = 1;
-    }
-    else
-        if(prec < 8 && nr_comp > 1)/* GRAY_ALPHA, RGB, RGB_ALPHA */
-        {
-            ushift = 8 - prec; dshift = 8 - ushift;
-            prec = 8; force8 = 1;
-        }
-
-    if(prec != 1 && prec != 2 && prec != 4 && prec != 8 && prec != 16)
-    {
-        fprintf(stderr,"imagetopng: can not create %s"
-                "\n\twrong bit_depth %d\n", write_idf, prec);
-        return fails;
-    }
-    writer = fopen(write_idf, "wb");
-
-    if(writer == NULL) return fails;
-
-    info = NULL; has_alpha = 0;
-
-    /* Create and initialize the png_struct with the desired error handler
- * functions.  If you want to use the default stderr and longjump method,
- * you can supply NULL for the last three parameters.  We also check that
- * the library version is compatible with the one used at compile time,
- * in case we are using dynamically linked libraries.  REQUIRED.
-*/
-    png = png_create_write_struct(PNG_LIBPNG_VER_STRING,
-                                  NULL, NULL, NULL);
-    /*png_voidp user_error_ptr, user_error_fn, user_warning_fn); */
-
-    if(png == NULL) goto fin;
-
-    /* Allocate/initialize the image information data.  REQUIRED
-*/
-    info = png_create_info_struct(png);
-
-    if(info == NULL) goto fin;
-
-    /* Set error handling.  REQUIRED if you are not supplying your own
- * error handling functions in the png_create_write_struct() call.
-*/
-    if(setjmp(png_jmpbuf(png))) goto fin;
-
-    /* I/O initialization functions is REQUIRED
-*/
-    png_init_io(png, writer);
-
-    /* Set the image information here.  Width and height are up to 2^31,
- * bit_depth is one of 1, 2, 4, 8, or 16, but valid values also depend on
- * the color_type selected. color_type is one of PNG_COLOR_TYPE_GRAY,
- * PNG_COLOR_TYPE_GRAY_ALPHA, PNG_COLOR_TYPE_PALETTE, PNG_COLOR_TYPE_RGB,
- * or PNG_COLOR_TYPE_RGB_ALPHA.  interlace is either PNG_INTERLACE_NONE or
- * PNG_INTERLACE_ADAM7, and the compression_type and filter_type MUST
- * currently be PNG_COMPRESSION_TYPE_BASE and PNG_FILTER_TYPE_BASE.
- * REQUIRED
- *
- * ERRORS:
- *
- * color_type == PNG_COLOR_TYPE_PALETTE && bit_depth > 8
- * color_type == PNG_COLOR_TYPE_RGB && bit_depth < 8
- * color_type == PNG_COLOR_TYPE_GRAY_ALPHA && bit_depth < 8
- * color_type == PNG_COLOR_TYPE_RGB_ALPHA) && bit_depth < 8
- *
-*/
-    png_set_compression_level(png, Z_BEST_COMPRESSION);
-
-    if(prec == 16) mask = 0xffff;
-    else
-        if(prec == 8) mask = 0x00ff;
-        else
-            if(prec == 4) mask = 0x000f;
-            else
-                if(prec == 2) mask = 0x0003;
-                else
-                    if(prec == 1) mask = 0x0001;
-
-    if(nr_comp >= 3
-            && image->comps[0].dx == image->comps[1].dx
-            && image->comps[1].dx == image->comps[2].dx
-            && image->comps[0].dy == image->comps[1].dy
-            && image->comps[1].dy == image->comps[2].dy
-            && image->comps[0].prec == image->comps[1].prec
-            && image->comps[1].prec == image->comps[2].prec)
-    {
-        int v;
-
-        has_alpha = (nr_comp > 3);
-
-        is16 = (prec == 16);
-
-        width = (int)image->comps[0].w;
-        height = (int)image->comps[0].h;
-
-        red = image->comps[0].data;
-        green = image->comps[1].data;
-        blue = image->comps[2].data;
-
-        sig_bit.red = sig_bit.green = sig_bit.blue = (png_byte)prec;
-
-        if(has_alpha)
-        {
-            sig_bit.alpha = (png_byte)prec;
-            alpha = image->comps[3].data;
-            color_type = PNG_COLOR_TYPE_RGB_ALPHA;
-            adjustA = (image->comps[3].sgnd ? 1 << (image->comps[3].prec - 1) : 0);
-        }
-        else
-        {
-            sig_bit.alpha = 0; alpha = NULL;
-            color_type = PNG_COLOR_TYPE_RGB;
-            adjustA = 0;
-        }
-        png_set_sBIT(png, info, &sig_bit);
-
-        png_set_IHDR(png, info, (png_uint_32)width, (png_uint_32)height, prec,
-                     color_type,
-                     PNG_INTERLACE_NONE,
-                     PNG_COMPRESSION_TYPE_BASE,  PNG_FILTER_TYPE_BASE);
-
-        png_set_gamma(png, 2.2, 1./2.2);
-        png_set_sRGB(png, info, PNG_sRGB_INTENT_PERCEPTUAL);
-        /*=============================*/
-        png_write_info(png, info);
-        /*=============================*/
-        if(prec < 8)
-        {
-            png_set_packing(png);
-        }
-// printf("%s:%d:sgnd(%d,%d,%d) w(%d) h(%d) alpha(%d)\n",__FILE__,__LINE__,
-//image->comps[0].sgnd,
-//image->comps[1].sgnd,image->comps[2].sgnd,width,height,has_alpha);
-
-        adjustR = (image->comps[0].sgnd ? 1 << (image->comps[0].prec - 1) : 0);
-        adjustG = (image->comps[1].sgnd ? 1 << (image->comps[1].prec - 1) : 0);
-        adjustB = (image->comps[2].sgnd ? 1 << (image->comps[2].prec - 1) : 0);
-
-        row_buf = (unsigned char*)malloc((size_t)width * (size_t)nr_comp * 2);
-
-        for(y = 0; y < height; ++y)
-        {
-            d = row_buf;
-
-            for(x = 0; x < width; ++x)
-            {
-                if(is16)
-                {
-                    v = *red + adjustR; ++red;
-		if(v > 65535) v = 65535; else if(v < 0) v = 0;
-
-                    if(force16) { v = (v<<ushift) + (v>>dshift); }
-
-                    *d++ = (unsigned char)(v>>8); *d++ = (unsigned char)v;
-
-                    v = *green + adjustG; ++green;
-		if(v > 65535) v = 65535; else if(v < 0) v = 0;
-
-                    if(force16) { v = (v<<ushift) + (v>>dshift); }
-
-                    *d++ = (unsigned char)(v>>8); *d++ = (unsigned char)v;
-
-                    v =  *blue + adjustB; ++blue;
-		if(v > 65535) v = 65535; else if(v < 0) v = 0;
-
-                    if(force16) { v = (v<<ushift) + (v>>dshift); }
-
-                    *d++ = (unsigned char)(v>>8); *d++ = (unsigned char)v;
-
-                    if(has_alpha)
-                    {
-                        v = *alpha + adjustA; ++alpha;
-		if(v > 65535) v = 65535; else if(v < 0) v = 0;
-
-                        if(force16) { v = (v<<ushift) + (v>>dshift); }
-
-                        *d++ = (unsigned char)(v>>8); *d++ = (unsigned char)v;
-                    }
-                    continue;
-                }/* if(is16) */
-
-                v = *red + adjustR; ++red;
-		if(v > 255) v = 255; else if(v < 0) v = 0;
-
-                if(force8) { v = (v<<ushift) + (v>>dshift); }
-
-                *d++ = (unsigned char)(v & mask);
-
-                v = *green + adjustG; ++green;
-		if(v > 255) v = 255; else if(v < 0) v = 0;
-
-                if(force8) { v = (v<<ushift) + (v>>dshift); }
-
-                *d++ = (unsigned char)(v & mask);
-
-                v = *blue + adjustB; ++blue;
-		if(v > 255) v = 255; else if(v < 0) v = 0;
-
-                if(force8) { v = (v<<ushift) + (v>>dshift); }
-
-                *d++ = (unsigned char)(v & mask);
-
-                if(has_alpha)
-                {
-                    v = *alpha + adjustA; ++alpha;
-		if(v > 255) v = 255; else if(v < 0) v = 0;
-
-                    if(force8) { v = (v<<ushift) + (v>>dshift); }
-
-                    *d++ = (unsigned char)(v & mask);
-                }
-            }	/* for(x) */
-
-            png_write_row(png, row_buf);
-
-        }	/* for(y) */
-        free(row_buf);
-
-    }/* nr_comp >= 3 */
-    else
-        if(nr_comp == 1 /* GRAY */
-                || (   nr_comp == 2 /* GRAY_ALPHA */
-                       && image->comps[0].dx == image->comps[1].dx
-                       && image->comps[0].dy == image->comps[1].dy
-                       && image->comps[0].prec == image->comps[1].prec))
-        {
-            int v;
-
-            red = image->comps[0].data;
-
-            sig_bit.gray = (png_byte)prec;
-            sig_bit.red = sig_bit.green = sig_bit.blue = sig_bit.alpha = 0;
-            alpha = NULL; adjustA = 0;
-            color_type = PNG_COLOR_TYPE_GRAY;
-
-            if(nr_comp == 2)
-            {
-                has_alpha = 1; sig_bit.alpha = (png_byte)prec;
-                alpha = image->comps[1].data;
-                color_type = PNG_COLOR_TYPE_GRAY_ALPHA;
-                adjustA = (image->comps[1].sgnd ? 1 << (image->comps[1].prec - 1) : 0);
-            }
-            width = (int)image->comps[0].w;
-            height = (int)image->comps[0].h;
-
-            png_set_IHDR(png, info, (png_uint_32)width, (png_uint_32)height, sig_bit.gray,
-                         color_type,
-                         PNG_INTERLACE_NONE,
-                         PNG_COMPRESSION_TYPE_BASE,  PNG_FILTER_TYPE_BASE);
-
-            png_set_sBIT(png, info, &sig_bit);
-
-            png_set_gamma(png, 2.2, 1./2.2);
-            png_set_sRGB(png, info, PNG_sRGB_INTENT_PERCEPTUAL);
-            /*=============================*/
-            png_write_info(png, info);
-            /*=============================*/
-            adjustR = (image->comps[0].sgnd ? 1 << (image->comps[0].prec - 1) : 0);
-
-            if(prec < 8)
-            {
-                png_set_packing(png);
-            }
-
-            if(prec > 8)
-            {
-                row_buf = (unsigned char*)
-                        malloc((size_t)width * (size_t)nr_comp * sizeof(unsigned short));
-
-                for(y = 0; y < height; ++y)
-                {
-                    d = row_buf;
-
-                    for(x = 0; x < width; ++x)
-                    {
-                        v = *red + adjustR; ++red;
-		if(v > 65535) v = 65535; else if(v < 0) v = 0;
-
-                        if(force16) { v = (v<<ushift) + (v>>dshift); }
-
-                        *d++ = (unsigned char)(v>>8); *d++ = (unsigned char)v;
-
-                        if(has_alpha)
-                        {
-                            v = *alpha++;
-		if(v > 65535) v = 65535; else if(v < 0) v = 0;
-
-                            if(force16) { v = (v<<ushift) + (v>>dshift); }
-
-                            *d++ = (unsigned char)(v>>8); *d++ = (unsigned char)v;
-                        }
-                    }/* for(x) */
-                    png_write_row(png, row_buf);
-
-                }	/* for(y) */
-                free(row_buf);
-            }
-            else /* prec <= 8 */
-            {
-                row_buf = (unsigned char*)calloc((size_t)width, (size_t)nr_comp * 2);
-
-                for(y = 0; y < height; ++y)
-                {
-                    d = row_buf;
-
-                    for(x = 0; x < width; ++x)
-                    {
-                        v = *red + adjustR; ++red;
-		if(v > 255) v = 255; else if(v < 0) v = 0;
-
-                        if(force8) { v = (v<<ushift) + (v>>dshift); }
-
-                        *d++ = (unsigned char)(v & mask);
-
-                        if(has_alpha)
-                        {
-                            v = *alpha + adjustA; ++alpha;
-		if(v > 255) v = 255; else if(v < 0) v = 0;
-
-                            if(force8) { v = (v<<ushift) + (v>>dshift); }
-
-                            *d++ = (unsigned char)(v & mask);
-                        }
-                    }/* for(x) */
-
-                    png_write_row(png, row_buf);
-
-                }	/* for(y) */
-                free(row_buf);
-            }
-        }
-        else
-        {
-            fprintf(stderr,"imagetopng: can not create %s\n",write_idf);
-            goto fin;
-        }
-    png_write_end(png, info);
-
-    fails = 0;
-
-fin:
-
-    if(png)
-    {
-        png_destroy_write_struct(&png, &info);
-    }
-    fclose(writer);
-
-    if(fails) remove(write_idf);
-
-    return fails;
-}/* imagetopng() */
-#endif /* OPJ_HAVE_LIBPNG */
